+++ src/style/emailVerify.module.css
@@ -116,7 +116,6 @@ body {
 }
 
 .resendButton {
-  margin-top: 20px;
   width: 100%;
   height: 58px;
   padding: 16px 12px;
@@ -140,9 +139,4 @@ body {
   border-radius: 4px;
   cursor: not-allowed;
   font-size: 16px;
-}
-
-.verifyButton:hover,
-.resendButton:hover {
-  
 }
